+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- printableValue and printableKey functions to be used in UI
+
 ## [5.3.0] - 2021-11-25
 ### Changed
 - update ethers `~5.5.1`